Jenkins.REST

Documentation

newtype Jenkins a

Constructors

Jenkins

Fields

unJenkins :: F JenkinsF a

data JenkinsF a where

JenkinsF terms

Constructors

Get :: Method Complete f -> (ByteString -> a) -> JenkinsF a

Post :: (forall f. Method Complete f) -> ByteString -> (ByteString -> a) -> JenkinsF a

Conc :: [Jenkins b] -> ([b] -> a) -> JenkinsF a

IO :: IO a -> JenkinsF a

With :: (forall m. Request m -> Request m) -> Jenkins b -> (b -> a) -> JenkinsF a

liftJ :: JenkinsF a -> Jenkins a

List JenkinsF term to the Jenkins language

get :: Method Complete f -> Jenkins ByteString

GET query

post :: (forall f. Method Complete f) -> ByteString -> Jenkins ()

POST query (with payload)

post_ :: (forall f. Method Complete f) -> Jenkins ()

POST query (without payload)

concurrently :: [Jenkins a] -> Jenkins [a]

Do a list of queries concurrently

io :: IO a -> Jenkins a

Lift arbitrary IO action

with :: (forall m. Request m -> Request m) -> Jenkins a -> Jenkins a

Make custom local changes to Request

Convenience

postXML :: (forall f. Method Complete f) -> Document -> Jenkins ()

POST job's config.xml (in xml-conduit format)

reload :: Jenkins ()

Reload jenkins configuration from disk

restart :: Jenkins ()

Restart jenkins

type Host = String

type Port = Int

type User = ByteString

type Password = ByteString

type APIToken = ByteString

data Settings

Jenkins settings

_jenkins_api_token might as well be user's password, Jenkins does not make any distinction between these concepts

Constructors

Settings

Fields

_jenkins_host :: Host

_jenkins_port :: Port

_jenkins_user :: User

_jenkins_api_token :: APIToken

jenkins_user :: Lens' Settings User

jenkins_port :: Lens' Settings Port

jenkins_host :: Lens' Settings Host

jenkins_api_token :: Lens' Settings APIToken

runJenkins :: Settings -> Jenkins a -> IO a

Communicate with Jenkins REST API

Throws HttpException and also all exceptions that could be thrown by embedded IO actions.

tryRunJenkins :: Exception e => Settings -> Jenkins a -> IO (Either e a)

Communicate with Jenkins REST API

Catches specified exceptions from API calls and embedded IO actions

 tryRunJenkins settings ≡ try settings . runJenkins

runJenkinsP :: Monad m => (JenkinsF (m a) -> m a) -> Jenkins a -> m a

Pure communication with Jenkins REST API

Obviously, It may not communicate with any Jenkins instance, since you can choose Identity for m

Nevertheless, it's useful for testing purposes and as a building block for runJenkins

runJenkinsIO :: Manager -> Jenkins a -> ReaderT (Request (ResourceT IO)) (ResourceT IO) a

jenkinsIO :: Manager -> JenkinsF (ReaderT (Request (ResourceT IO)) (ResourceT IO) a) -> ReaderT (Request (ResourceT IO)) (ResourceT IO) a

module Jenkins.REST.Lens

module Jenkins.REST.Method

data Request m

All information on how to connect to a host and what should be sent in the HTTP request.

If you simply wish to download from a URL, see parseUrl.

The constructor for this data type is not exposed. Instead, you should use either the def method to retrieve a default instance, or parseUrl to construct from a URL, and then use the records below to make modifications. This approach allows http-conduit to add configuration options without breaking backwards compatibility.

For example, to construct a POST request, you could do something like:

 initReq <- parseUrl "http://www.example.com/path"
 let req = initReq
             { method = "POST"
             }

For more information, please see http://www.yesodweb.com/book/settings-types.

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.