Jenkins.REST.Internal

Documentation

newtype Jenkins a

Jenkins REST API composable queries

Constructors

Jenkins

Fields

unJenkins :: F JenkinsF a

data JenkinsF a where

JenkinsF terms

Constructors

Get :: Method Complete f -> (ByteString -> a) -> JenkinsF a

Post :: (forall f. Method Complete f) -> ByteString -> (ByteString -> a) -> JenkinsF a

Conc :: Jenkins a -> Jenkins b -> (a -> b -> c) -> JenkinsF c

IO :: IO a -> JenkinsF a

With :: (Request -> Request) -> Jenkins b -> (b -> a) -> JenkinsF a

Dcon :: JenkinsF a

liftJ :: JenkinsF a -> Jenkins a

Lift JenkinsF query to the Jenkins query language

runJenkins :: ConnectInfo -> Jenkins a -> IO (Either HttpException (Maybe a))

Query Jenkins REST API

Successful result can be either Just val or Nothing in case of premature disconnect (see reload, restart, or forceRestart)

Catches HttpExceptions thrown by http-conduit, but does not catch exceptions thrown by embedded IO actions

iterJenkinsIO :: Manager -> Jenkins a -> MaybeT (ReaderT Request (ResourceT IO)) a

Interpret JenkinsF AST in IO

iterJenkins :: Monad m => (JenkinsF (m a) -> m a) -> Jenkins a -> m a

Tear down JenkinsF AST with a JenkinsF-algebra

interpreter :: Manager -> JenkinsF (MaybeT (ReaderT Request (ResourceT IO)) a) -> MaybeT (ReaderT Request (ResourceT IO)) a

JenkinsF AST interpreter

data ConnectInfo

Jenkins connection settings

_jenkinsApiToken may be user's password, Jenkins does not make any distinction between these concepts

Constructors

ConnectInfo

Fields

_jenkinsUrl :: String

Jenkins URL, e.g. http://example.com/jenkins

_jenkinsPort :: Int

Jenkins port, e.g. 8080

_jenkinsUser :: ByteString

Jenkins user, e.g. jenkins

_jenkinsApiToken :: ByteString

Jenkins user API token

defaultConnectInfo :: ConnectInfo

Default Jenkins connection settings

 defaultConnectInfo = ConnectInfo
   { _jenkinsUrl      = "http://example.com/jenkins"
   , _jenkinsPort     = 8080
   , _jenkinsUser     = "jenkins"
   , _jenkinsApiToken = ""
   }

jenkinsUrl :: Lens' ConnectInfo String

A lens into Jenkins URL

jenkinsPort :: Lens' ConnectInfo Int

A lens into Jenkins port

jenkinsUser :: Lens' ConnectInfo ByteString

A lens into Jenkins user

jenkinsApiToken :: Lens' ConnectInfo ByteString

A lens into Jenkins user API token

jenkinsPassword :: Lens' ConnectInfo ByteString

A lens into Jenkins password

 jenkinsPassword = jenkinsApiToken