Jenkins.REST.Lens

Description

Lenses for Conduit types

Request lenses

method :: Lens' (Request m) Method

HTTP request method, eg GET, POST.

secure :: Lens' (Request m) Bool

Whether to use HTTPS (ie, SSL).

host :: Lens' (Request m) ByteString

port :: Lens' (Request m) Int

path :: Lens' (Request m) ByteString

Everything from the host to the query string.

queryString :: Lens' (Request m) ByteString

requestBody :: Lens' (Request m) (RequestBody m)

requestHeaders :: Lens' (Request m) RequestHeaders

redirectCount :: Lens' (Request m) Int

How many redirects to follow when getting a resource. 0 means follow no redirects. Default value: 10.

checkStatus :: Lens' (Request m) (Status -> ResponseHeaders -> CookieJar -> Maybe SomeException)

Check the status code. Note that this will run after all redirects are performed. Default: return a StatusCodeException on non-2XX responses.

responseTimeout :: Lens' (Request m) (Maybe Int)

Number of microseconds to wait for a response. If Nothing, will wait indefinitely. Default: 5 seconds.