Jenkins.Rest.Method

Description

Jenkins REST API methods

Construct URLs

Path

text :: Text -> Method Complete f

Use a string as an URI segment

>>> pp (text "foo")
"foo"

Note: with -XOverloadedStrings extension enabled it's possible to use string literals as segments of the Jenkins API method URL

>>> pp' "foo"
"foo"

Note: don't put / in the string literal unless you want it URL-encoded, use (-/-) instead

>>> pp' "foo/bar"
"foo%2Fbar"

int :: Int -> Method Complete f

Use an integer as an URI segment

>>> pp (int 4)
"4"

(-/-) :: Method Complete f -> Method Complete f -> Method Complete f infixr 5

Combine two paths

>>> pp ("foo" -/- "bar" -/- "baz")
"foo/bar/baz"

Query

(-=-) :: Text -> Text -> Method Query f infix 7

Make a key-value pair

>>> pp ("foo" -=- "bar")
"foo=bar"

(-&-) :: Method Query f -> Method Query f -> Method Query f infixr 5

Create the union of two queries

>>> pp ("foo" -=- "bar" -&- "baz")
"foo=bar&baz"

query :: [(Text, Maybe Text)] -> Method Query f

Take a list of key-value pairs and render them as a query

>>> pp (query [("foo", Nothing), ("bar", Just "baz"), ("quux", Nothing)])
"foo&bar=baz&quux"

>>> pp (query [])
""

Put together the segments and the query

(-?-) :: Method Complete f -> Method Query f -> Method Complete f infix 1

Put path and query together

>>> pp ("qux" -/- "quux" -?- "foo" -=- "bar" -&- "baz")
"qux/quux?foo=bar&baz"

Format

data Formatter g

Formatters know how to append the "api/$format" string to the method URL

json :: Formatter Json

Append the JSON formatting request to the method URL

>>> format json "foo"
"foo/api/json"

xml :: Formatter Xml

Append the XML formatting request to the method URL

>>> format xml "foo"
"foo/api/xml"

python :: Formatter Python

Append the Python formatting request to the method URL

>>> format python "foo"
"foo/api/python"

plain :: Formatter f

The formatter that does exactly nothing

>>> format plain "foo"
"foo"

Shortcuts

job :: Text -> Method Complete f

Job data

>>> format json (job "name")
"job/name/api/json"

>>> pp (job "name" -/- "config.xml")
"job/name/config.xml"

build :: Text -> Int -> Method Complete f

Job build data

>>> format json (build "name" 4)
"job/name/4/api/json"

view :: Text -> Method Complete f

View data

>>> format xml (view "name")
"view/name/api/xml"

queue :: Method Complete f

Build queue data

>>> format python queue
"queue/api/python"

overallLoad :: Method Complete f

Server statistics

>>> format xml overallLoad
"overallLoad/api/xml"

computer :: Method Complete f

Nodes data

>>> format python computer
"computer/api/python"

Types

data Method :: Type -> Format -> *

Jenkins RESTFul API method encoding

data Type

Method types

Constructors

Query

Complete

data Format

Response formats

Constructors

Json

Xml

Python