Jenkins.REST

Description

Interface to Jenkins REST API

Types

data Jenkins a

Jenkins REST API composable queries

data Settings

Jenkins settings

_jenkins_api_token might as well be the user's password, Jenkins does not make any distinction between these concepts

Constructors

Settings

Fields

_jenkins_host :: Host

_jenkins_port :: Port

_jenkins_user :: User

_jenkins_api_token :: APIToken

newtype Host

Jenkins host address

Constructors

Host

Fields

unHost :: String

newtype Port

Jenkins port

Constructors

Port

Fields

unPort :: Int

newtype User

Jenkins user

Constructors

User

Fields

unUser :: ByteString

newtype APIToken

Jenkins user API token

Constructors

APIToken

Fields

unAPIToken :: ByteString

type Password = APIToken

Jenkins user password

data Request m

All information on how to connect to a host and what should be sent in the HTTP request.

If you simply wish to download from a URL, see parseUrl.

The constructor for this data type is not exposed. Instead, you should use either the def method to retrieve a default instance, or parseUrl to construct from a URL, and then use the records below to make modifications. This approach allows http-conduit to add configuration options without breaking backwards compatibility.

For example, to construct a POST request, you could do something like:

 initReq <- parseUrl "http://www.example.com/path"
 let req = initReq
             { method = "POST"
             }

For more information, please see http://www.yesodweb.com/book/settings-types.

Lenses and prisms

jenkins_host :: Lens' Settings Host

jenkins_port :: Lens' Settings Port

jenkins_user :: Lens' Settings User

jenkins_api_token :: Lens' Settings APIToken

jenkins_password :: Lens' Settings Password

_Host :: Prism' Host String

_Port :: Prism' Port Int

_User :: Prism' User ByteString

_APIToken :: Prism' APIToken ByteString

_Password :: Prism' Password ByteString

Jenkins queries construction

get :: Method Complete f -> Jenkins ByteString

GET query

post :: (forall f. Method Complete f) -> ByteString -> Jenkins ()

POST query (with payload)

post_ :: (forall f. Method Complete f) -> Jenkins ()

POST query (without payload)

concurrently :: Jenkins a -> Jenkins b -> Jenkins (a, b)

Do both queries concurrently

io :: IO a -> Jenkins a

Lift arbitrary IO action

with :: (forall m. Request m -> Request m) -> Jenkins a -> Jenkins a

Make custom local changes to Request

module Jenkins.REST.Method

Little helpers

concurrentlys :: [Jenkins a] -> Jenkins [a]

Do a list of queries concurrently

concurrentlys_ :: [Jenkins a] -> Jenkins ()

Do a list of queries concurrently

postXML :: (forall f. Method Complete f) -> Document -> Jenkins ()

POST job's config.xml (in xml-conduit format)

reload :: Jenkins ()

Reload jenkins configuration from disk

restart :: Jenkins ()

Restart jenkins

Jenkins queries execution

runJenkins :: Settings -> Jenkins a -> IO a

Communicate with Jenkins REST API

Throws HttpException and also all exceptions that could be thrown by embedded IO actions.

tryRunJenkins :: Exception e => Settings -> Jenkins a -> IO (Either e a)

Communicate with Jenkins REST API

Catches specified exceptions from API calls and embedded IO actions

 tryRunJenkins settings ≡ try settings . runJenkins

runJenkinsP :: Monad m => (JenkinsF (m a) -> m a) -> Jenkins a -> m a

Generic Jenkins REST API queries interpreter

Particularly useful for testing (with m ≡ Identity)

Usable http-conduit Request type API

module Jenkins.REST.Lens