Jenkins.REST.Internal

Documentation

newtype Jenkins a

Jenkins query EDSL abstract type

Constructors

Jenkins

Fields

unJenkins :: F JenkinsF a

data JenkinsF a where

JenkinsF terms

Constructors

Get :: Method Complete f -> (ByteString -> a) -> JenkinsF a

Post :: (forall f. Method Complete f) -> ByteString -> (ByteString -> a) -> JenkinsF a

Conc :: Jenkins a -> Jenkins b -> (a -> b -> c) -> JenkinsF c

IO :: IO a -> JenkinsF a

With :: (forall m. Request m -> Request m) -> Jenkins b -> (b -> a) -> JenkinsF a

liftJ :: JenkinsF a -> Jenkins a

List JenkinsF term to the Jenkins language

runJenkinsIO :: Manager -> Jenkins a -> ReaderT (Request (ResourceT IO)) (ResourceT IO) a

runJenkinsP :: Monad m => (JenkinsF (m a) -> m a) -> Jenkins a -> m a

Possibly Pure communication with Jenkins REST API

Obviously, It may not communicate with any Jenkins instance, since you can choose Identity for m

Nevertheless, it's useful for testing purposes and as a building block for runJenkins

jenkinsIO :: Manager -> JenkinsF (ReaderT (Request (ResourceT IO)) (ResourceT IO) a) -> ReaderT (Request (ResourceT IO)) (ResourceT IO) a