Jenkins.Rest

Description

Jenkins REST API interface

Query Jenkins

runJenkins :: HasConnectInfo t => t -> Jenkins a -> IO (Result JenkinsException a)

Query Jenkins API using Jenkins description

Successful result is either Disconnect or Result v

If HttpException was thrown by http-conduit, runJenkins catches it and wraps in Error. Other exceptions are not catched

data Result e v

The result of Jenkins REST API queries

Constructors

Error e
Exception e was thrown while querying Jenkins

Disconnect
The client was explicitly disconnected

Result v
Querying successfully finished the with value v

data Jenkins a

Jenkins REST API query sequence description

class HasConnectInfo t where

Convenience class aimed at elimination of long chains of lenses to access jenkins connection configuration

For example, if you have a configuration record in your application:

data Config = Config
  { ...
  , _jenkinsConnectInfo :: ConnectInfo
  , ...
  }

you can make it an instance of HasConnectInfo:

instance HasConnectInfo Config where
  connectInfo f x = (p -> x { _jenkinsConnectInfo = p }) <$> f (_jenkinsConnectInfo x)

and then use e.g. view jenkinsUrl config to get the url part of the jenkins connection

Minimal complete definition

connectInfo

Methods

connectInfo :: Lens' t ConnectInfo

jenkinsUrl :: HasConnectInfo t => Lens' t String

A lens into Jenkins URL

jenkinsUser :: HasConnectInfo t => Lens' t Text

A lens into username to access the Jenkins instance with

jenkinsApiToken :: HasConnectInfo t => Lens' t Text

A lens into user API token or password

data ConnectInfo

Jenkins connection settings

defaultConnectInfo :: ConnectInfo

Default Jenkins connection settings

defaultConnectInfo = ConnectInfo
  { _jenkinsUrl      = "http://example.com/jenkins"
  , _jenkinsUser     = "jenkins"
  , _jenkinsApiToken = ""
  }

Combinators

get :: Method Complete f -> Jenkins ByteString

GET query

While the return type is a lazy bytestring, the entire response sits in memory anyway: lazy I/O is not used

getS :: (MonadCatch m, MonadResource m) => Method Complete f -> Jenkins (m (ResumableSource m ByteString))

GET query

If you don't close the source eventually (either explicitly with closeResumableSource or implicitly by reading from it) it will leak a socket.

post :: (forall f. Method Complete f) -> ByteString -> Jenkins ()

POST query (with a payload)

post_ :: (forall f. Method Complete f) -> Jenkins ()

POST query (without payload)

concurrently :: Jenkins a -> Jenkins b -> Jenkins (a, b)

Do both queries concurrently

orElse :: Jenkins a -> Jenkins a -> Jenkins a

orElse a b runs a and only runs b if a has thrown a JenkinsException

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

with :: (Request -> Request) -> Jenkins a -> Jenkins a

Make local changes to the Request

Method

module Jenkins.Rest.Method

Convenience

postXML :: (forall f. Method Complete f) -> Document -> Jenkins ()

POST job's config.xml (or any other xml, really) in xml-conduit format

traverseC :: (a -> Jenkins b) -> [a] -> Jenkins [b]

Make a bunch of queries concurrently

traverseC_ :: Foldable f => (a -> Jenkins b) -> f a -> Jenkins ()

Make a bunch of queries concurrently ignoring their results

reload :: Jenkins a

Reload jenkins configuration from disk

Calls /reload and disconnects

restart :: Jenkins a

Restart jenkins safely

Calls /safeRestart and disconnects

/safeRestart allows all running jobs to complete

forceRestart :: Jenkins a

Force jenkins to restart without waiting for running jobs to finish

Calls /restart and disconnects

Optics

_Error :: Prism (Result e a) (Result e' a) e e'

A prism into Jenkins error

_Disconnect :: Prism' (Result e a) ()

A prism into disconnect

_Result :: Prism (Result e a) (Result e b) a b

A prism into result

newtype JenkinsException

Constructors

JenkinsHttpException HttpException

Reexports

data Request :: *